Fixed transcribe_results.sh to only copy new files 🐛

+++ data/scripts/transcribe_results.sh
@@ -16,7 +16,7 @@ TRIBUS_RESULT_PATH="result/transcription.trs"
 TRIBUS_RESULT_CTM="/ctm"
 TRIBUS_RESULT_PCTM="/pctm"
 
-if [ -d "${TRANSCRIPTIONS_RESULTS_PATH}" ]; then rm -rf "${TRANSCRIPTIONS_RESULTS_PATH}"; fi
+if [ ! -d "${TRANSCRIPTIONS_RESULTS_PATH}" ]; then mkdir "${TRANSCRIPTIONS_RESULTS_PATH}"; fi
 
 echo "Extracting results from Controls ..."
 for taskFolder in `ls ${CONTROL_TRANSCRIPTIONS_PATH}${TRIBUS_EXPORT_PATH}`; do
@@ -51,9 +51,10 @@ for taskFolder in `ls ${CONTROL_TRANSCRIPTIONS_PATH}${TRIBUS_EXPORT_PATH}`; do
         #echo "Filename: ${filename}"
 
 	if [ ! -d "${dir_out_path}" ]; then mkdir -p "${dir_out_path}"; fi
-	cp ${file_in_trans_path} ${file_out_trans_path}
-	cp ${file_in_ctm_path} ${file_out_ctm_path}
-	cp ${file_in_pctm_path} ${file_out_pctm_path}
+
+        if [ ! -f "${file_out_trans_path}" ]; then cp ${file_in_trans_path} ${file_out_trans_path}; fi
+        if [ ! -f "${file_out_ctm_path}" ];cp ${file_in_ctm_path} ${file_out_ctm_path}; fi
+        if [ ! -f "${file_out_pctm_path}" ];cp ${file_in_pctm_path} ${file_out_pctm_path}; fi
 done
 
 echo "Extracting results from Psychosis ..."
@@ -89,9 +90,10 @@ for taskFolder in `ls ${PSYCHOSIS_TRANSCRIPTIONS_PATH}${TRIBUS_EXPORT_PATH}`; do
         #echo "Filename: ${filename}"
 
         if [ ! -d "${dir_out_path}" ]; then mkdir -p "${dir_out_path}"; fi
-        cp ${file_in_trans_path} ${file_out_trans_path}
-        cp ${file_in_ctm_path} ${file_out_ctm_path}
-        cp ${file_in_pctm_path} ${file_out_pctm_path}
+
+        if [ ! -f "${file_out_trans_path}" ]; then cp ${file_in_trans_path} ${file_out_trans_path}; fi
+        if [ ! -f "${file_out_ctm_path}" ];cp ${file_in_ctm_path} ${file_out_ctm_path}; fi
+        if [ ! -f "${file_out_pctm_path}" ];cp ${file_in_pctm_path} ${file_out_pctm_path}; fi
 done
 
 echo "Extracting results from Bipolars ..."
@@ -127,7 +129,8 @@ for taskFolder in `ls ${BIPOLARS_TRANSCRIPTIONS_PATH}${TRIBUS_EXPORT_PATH}`; do
         #echo "Filename: ${filename}"
 
         if [ ! -d "${dir_out_path}" ]; then mkdir -p "${dir_out_path}"; fi
-        cp ${file_in_trans_path} ${file_out_trans_path}
-        cp ${file_in_ctm_path} ${file_out_ctm_path}
-        cp ${file_in_pctm_path} ${file_out_pctm_path}
+
+        if [ ! -f "${file_out_trans_path}" ]; then cp ${file_in_trans_path} ${file_out_trans_path}; fi
+        if [ ! -f "${file_out_ctm_path}" ];cp ${file_in_ctm_path} ${file_out_ctm_path}; fi
+        if [ ! -f "${file_out_pctm_path}" ];cp ${file_in_pctm_path} ${file_out_pctm_path}; fi
 done
